# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "School Lunch Battle Internal API",
    "version": "1.0.0"
  },
  "paths": {
    "/health/live": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/health/ready": {
      "get": {
        "responses": {
          "200": {
            "description": "Ready"
          }
        }
      }
    },
    "/api/v1/schools": {
      "get": {
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "School search result"
          },
          "422": {
            "description": "Validation error"
          },
          "429": {
            "description": "Rate limited"
          },
          "502": {
            "description": "NEIS error"
          },
          "503": {
            "description": "NEIS unavailable"
          }
        }
      }
    },
    "/api/v1/meals": {
      "get": {
        "parameters": [
          {
            "name": "officeCode",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "schoolCode",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Meal lookup result"
          },
          "422": {
            "description": "Validation error"
          },
          "429": {
            "description": "Rate limited"
          },
          "502": {
            "description": "NEIS error"
          },
          "503": {
            "description": "NEIS unavailable"
          }
        }
      }
    }
  }
}